requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==22.1
python-decouple==3.8
Pillow==11.3.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
urllib3==1.26.18
cryptography==41.0.7
pycryptodome==3.19.0
django-redis==5.4.0
django-cacheops==7.2
django-silk==5.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
python-decouple==3.8
Pillow==11.3.0
